token save, update

+++ src/main/java/com/sh/shpay/domain/openbanking/openbanking/api/dto/res/OpenBankingUserRefreshTokenResponseDto.java
@@ -14,7 +14,7 @@ public class OpenBankingUserRefreshTokenResponseDto {
 
     private String access_token;
     private String token_type;
-    private int expires_in;
+    private Long expires_in;
     private String refresh_token;
     private String scope;
     private String user_seq_no;
